backend/requirements.txt
# FastAPI and web server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.0

# PydanticAI and LLM
pydantic-ai==0.6.2
pydantic-settings==2.5.2
openai==1.99.8

# PDF processing
pdfplumber==0.11.4
pypdf==4.3.1
pdf2image==1.17.0

# Data validation and utilities
pydantic>=2.10.0
typing-extensions==4.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.27.2

# Code quality
ruff==0.1.6
mypy==1.7.1

# Additional utilities (removed non-existent packages)